drive: move normalizeShareName into pkg drive and make func public
This change makes the normalizeShareName function public, so it can be
used for validation in control.

Updates tailscale/corp#16827
